requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
